Orchestrator/agent2/requirements.txt
langgraph
langchain
langchain-google-genai
mysql-connector-python
python-dotenv
